+++ source/isaaclab/docs/CHANGELOG.rst
@@ -10,7 +10,7 @@ Added
 
 * Added :class:`~isaaclab.sim.views.BaseXformPrimView` abstract base class that defines
   the common interface for backend-specific ``XformPrimView`` implementations.
-* Added :class:`~isaaclab.sim.views.XformPrimViewFactory` to instantiate the correct
+* Added :class:`~isaaclab.sim.views.XformPrimView` factory to instantiate the correct
   backend-specific ``XformPrimView`` based on the active simulation backend.
 
 Changed
@@ -19,7 +19,7 @@ Changed
 * Refactored :class:`~isaaclab.sim.views.XformPrimView` to delegate backend-specific
   logic to :class:`~isaaclab_physx.sim.views.XformPrimView` and
   :class:`~isaaclab_newton.sim.views.XformPrimView`. The public API is unchanged;
-  use :class:`~isaaclab.sim.views.XformPrimViewFactory` for backend-aware instantiation.
+  use :class:`~isaaclab.sim.views.XformPrimView` for backend-aware instantiation.
 
 * Added release version to
   :class:`~isaaclab.test.benchmark.recorders.VersionInfoRecorder` output.
